+++ package.json
@@ -13,21 +13,21 @@
   },
   "dependencies": {
     "@astrojs/react": "^2.2.1",
-    "@types/react": "^18.0.21",
-    "@types/react-dom": "^18.0.6",
-    "astro": "^2.9.3",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.3",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
+    "astro": "^2.9.3",
     "axios": "^0.21.1",
     "bootstrap": "^4.6.0",
     "gh-pages": "^3.1.0",
     "js-cookie": "^2.2.1",
     "node-sass": "^8.0.0",
+    "react": "^18.2.0",
     "react-bootstrap": "^1.5.2",
     "react-color": "^2.19.3",
+    "react-dom": "^18.2.0",
     "react-google-login": "^5.2.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
@@ -54,6 +54,7 @@
   },
   "devDependencies": {
     "autoprefixer": "^9.8.6",
-    "postcss": "^7.0.35"
+    "postcss": "^7.0.35",
+    "sass": "^1.64.1"
   }
 }
