+++ src/components/PlanDeEstudio/BarrasDeProgreso/BarrasDeProgreso.jsx
