+++ src/components/PlanDeEstudio/BarrasDeProgreso/BarrasDeProgreso.test.jsx
