+++ src/components/PlanDeEstudio/BotonesDeColor/BotonesDeColor.jsx
