+++ src/components/PlanDeEstudio/BotonesDeColor/BotonesDeColor.test.jsx
