+++ src/components/Profile/Profile.jsx
@@ -4,8 +4,8 @@ import axios from "axios";
 import { useToasts } from 'react-toast-notifications';
 
 import Planificado from "./Planificado";
-import { BACKEND_URL, toQueryString } from "../utils"; 
-import { UserContext } from "../../../react_src/context";
+import { BACKEND_URL, toQueryString } from "../../../utils/utils"; 
+import { UserContext } from "../../context";
 
 async function fetchPlanificados(usuario, setPlanes, addToast) {
   const query = toQueryString({ usuario });
