+++ src/components/shared/Header.jsx
@@ -2,10 +2,10 @@ import React, { useContext } from 'react'
 import { Navbar, Image, Nav, Button, Container } from 'react-bootstrap'
 import { GoogleLogin } from "react-google-login";
 
-import { PUBLIC_URL } from './../utils';
+import { PUBLIC_URL } from '../../../utils/utils';
 import { login, logout, G_CLIENT_ID } from "../../../utils/auth";
 import logoutIcon from "./logout_white_24dp.svg";
-import { UserContext } from "../../../react_src/context";
+import { UserContext } from "../../context";
 
 /**
  * Parte superior, contiene controles de manejo de sesión y perfil
