+++ src/components/shared/Header.test.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from '@testing-library/react';
 
 import Header from "./Header";
-import { UserContext } from "../../../react_src/context";
+import { UserContext } from "../../context";
 
 it("renderiza header cuando no se ha iniciado sesión", () => {
   const { getByText } = render(<Header/>);
