+++ src/pages/index.astro
@@ -1,11 +1,11 @@
 ---
 
-import App from "../components/App"
+import IndexComponent from "../components/IndexComponent"
 
 ---
 
 <html>
   <body>
-    <App />
+    <IndexComponent />
   </body>
 </html>
